DEX-8606 Update public Cloudera Airflow plugins.
