+++ .gitignore
@@ -1,5 +1,13 @@
-.DS_Store
 
+.idea/
+venv/
+.tox/
+dist/
+*.egg-info
+.eggs/
+build/
+
+.DS_Store
 .pytest_cache/
 __pycache__/
 *.pyc
