+++ README.md
@@ -1,6 +1,35 @@
-# cloudera-airflow-plugins
+![Cloudera logo](docs/images/Cloudera.png)
 
-Collection of Airflow related adds-on/plugins specific to Cloudera.
+Collection of Airflow related adds-on specific to Cloudera.
 
-For getting started and more information about a specific component, please refer to the README in the corresponding section:
-* [Cloudera Airflow Provider](cloudera_airflow_provider)
+## Cloudera Airflow provider.
+
+Data orchestration with Apache Airflow securely at scale to increase efficiency and accelerate time to value.
+
+- [Overview](/cloudera_airflow_provider#overview)
+
+- [Getting started](/cloudera_airflow_provider#getting-started)
+    
+    - [Installation](/cloudera_airflow_provider#installation)
+    
+    - [Setting up Airflow Connections](/cloudera_airflow_provider#setting-up-the-airflow-connections)
+    
+        - [Cloudera Data Engineering connection](/cloudera_airflow_provider#cloudera-data-engineering-connection)
+        
+        - [Hive CLI connection to Cloudera Data Warehouse](/cloudera_airflow_provider#hive-cli-connection-to-cloudera-data-warehouse)
+        
+- [Usage details](/cloudera_airflow_provider#usage-details)
+
+    - [CdeRunJobOperator](/cloudera_airflow_provider#cderunjoboperator)
+    
+    - [CdwExecuteQueryOperator](/cloudera_airflow_provider#cdwexecutequeryoperator)
+        
+        - [CDW operator preconditions](docs/cdw-operator-preconditions.md)
+    
+- [Next steps](/cloudera_airflow_provider#next-steps)
+
+    - [CDE Concepts](https://docs.cloudera.com/data-engineering/cloud/cli-access/topics/cde-cli-concepts.html)
+    
+    - [Example DAGs](/docs/examples)
+    
+    - [Integrate operators into your pipelines](https://docs.cloudera.com/data-engineering/cloud/orchestrate-workflows/topics/cde-airflow-dag-pipeline.html)
