+++ cloudera_airflow_provider/cloudera/cdp/airflow/operators/cde_operator.py
@@ -32,14 +32,16 @@
 #  RELATED TO LOST REVENUE, LOST PROFITS, LOSS OF INCOME, LOSS OF
 #  BUSINESS ADVANTAGE OR UNAVAILABILITY, OR LOSS OR CORRUPTION OF
 #  DATA.
-"""This module is deprecated. Please use :mod:`cloudera.airflow.providers.operators.cde_operator`."""
+"""This module is deprecated. Please use :mod:`cloudera.airflow.providers.operators.cde`."""
 
 import warnings
 
-from cloudera.airflow.providers.operators.cde_operator import CdeRunJobOperator as CDEJobRunOperator # pylint: disable=unused-import
+from cloudera.airflow.providers.operators.cde import (  # pylint: disable=unused-import     # noqa:F401
+    CdeRunJobOperator as CDEJobRunOperator,
+)
 
 warnings.warn(
-    "This module is deprecated. Please use `cloudera.airflow.providers.operators.cde_operator`.",
+    "This module is deprecated. Please use `cloudera.airflow.providers.operators.cde`.",
     DeprecationWarning,
     stacklevel=2,
 )
