+++ cloudera_airflow_provider/cloudera/cdp/airflow/operators/cdw_operator.py
@@ -32,14 +32,16 @@
 #  RELATED TO LOST REVENUE, LOST PROFITS, LOSS OF INCOME, LOSS OF
 #  BUSINESS ADVANTAGE OR UNAVAILABILITY, OR LOSS OR CORRUPTION OF
 #  DATA.
-"""This module is deprecated. Please use :mod:`cloudera.airflow.providers.operators.cdw_operator`."""
+"""This module is deprecated. Please use :mod:`cloudera.airflow.providers.operators.cdw."""
 
 import warnings
 
-from cloudera.airflow.providers.operators.cdw_operator import CdwExecuteQueryOperator as CDWOperator # pylint: disable=unused-import
+from cloudera.airflow.providers.operators.cdw import (  # pylint: disable=unused-import     # noqa:F401
+    CdwExecuteQueryOperator as CDWOperator,
+)
 
 warnings.warn(
-    "This module is deprecated. Please use `cloudera.airflow.providers.operators.cdw_operator`.",
+    "This module is deprecated. Please use `cloudera.airflow.providers.operators.cdw`.",
     DeprecationWarning,
     stacklevel=2,
 )
