+++ cloudera_airflow_provider/setup.cfg
@@ -1,23 +1,34 @@
 [metadata]
 name = cloudera-airflow-provider
-version = 1.0.2
+version = 2.0.1
 author= Cloudera
 description= Cloudera CDP Airflow Plugins
 
-
 [options]
-packages=find:
-python_requires = >=3.6
+packages = find_namespace:
+python_requires = >=3.7
 install_requires=
   apache-airflow
-  pure25519
+  apache-airflow-providers-apache-hive
   # Minimum crypto version for handling CVE-2020-36242
   cryptography>=3.3.2
   tenacity
   requests
-  pathlib2
+  cloudera-cde-sdk>=1.0.1
 
 [options.entry_points]
 # Airflow 2 provider entrypoint
 apache_airflow_provider=
-  provider_info=cloudera:get_provider_info
+  provider_info=cloudera.cdp.airflow:get_provider_info
+
+[options.packages.find]
+include =
+  cloudera*
+
+[isort]
+line_length=110
+combine_as_imports = true
+default_section = THIRDPARTY
+known_first_party=airflow,tests
+skip=build,.tox,venv
+profile = black
