+++ cloudera_airflow_provider/setup.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
 
 import setuptools
+
 setuptools.setup()
