+++ cloudera_airflow_provider/tests/providers/cloudera/__init__.py
@@ -32,16 +32,3 @@
 #  RELATED TO LOST REVENUE, LOST PROFITS, LOSS OF INCOME, LOSS OF
 #  BUSINESS ADVANTAGE OR UNAVAILABILITY, OR LOSS OR CORRUPTION OF
 #  DATA.
-
-from cloudera.airflow.providers.hooks.cdw_hook import CdwHiveMetastoreHook
-
-
-def test_csv_parse():
-    """
-    This is just simple validation test for csv reader. The variable beeline_output
-    contains a sample response which comes from hive in case of --outputformat=csv2.
-    """
-    beeline_output = "db_name,tbl_name,part_name\n" "default,test_part,dt=1"
-    result_list = CdwHiveMetastoreHook.parse_csv_lines(beeline_output)
-
-    assert len(result_list) == 2, result_list
