+++ docs/examples/README.md
@@ -1,0 +1,9 @@
+# Cloudera Airflow plugins example DAGs.
+
+This directory contains example Directed Acyclic Graphs for Apache Airflow.
+
+They are:
+
+- [CdeRunJobOperator](cde_operator_example.py) example DAG.
+
+- [CdwExecuteQueryOperator](cdw_operator_example.py) example DAG.
